refactor(gpui): split body.rs, slim log/mod.rs, measure mono advance
Three threads bundled here because they accumulated on the same
working copy without intermediate `jj new`s — process slip on my
side, recorded honestly:

1. Split diff/diff_view/body.rs (was 353 lines):
   - unified_body.rs — unified-diff renderer
   - sbs_body.rs — sbs renderer + SbsContentArgs + sbs_content_list
   - mouse.rs — bounds_capture + pixel_to_col helpers shared by both

2. Slim shell/gpui/src/log/mod.rs (was 223 lines, now 20):
   - view.rs — LogView struct + impl, ActivePane / ColumnDrag /
     DragTarget enums, layout constants, PanelBoundsSlot, Focusable
   - window.rs — open_repo_window
   mod.rs is now a thin index of pub mod / pub use re-exports.

3. Replace MONO_GLYPH_WIDTH = 7.2 with measured advance via gpui's
   text_system::ch_advance, fixes the trailing-highlight bug:
   - app/fonts.rs: mono_advance(cx, size) helper
   - selection_overlay computes width = (end - start) * advance
     instead of anchoring right edge to the parent panel; cols.end
     is already clamped via DiffSelection::col_range_for, so each
     line's highlight stops at its own EOL — matches VSCode /
     GitHub Desktop, fixes the bug from screenshot feedback.
   - pixel_to_col + content_row + body fns thread the advance.
