+++ shell/gpui/src/app/fonts.rs
@@ -1,6 +1,7 @@
 use std::sync::OnceLock;
 
 use font_kit::source::SystemSource;
+use gpui::{App, Pixels, font, px};
 
 const MONO_CANDIDATES: &[&str] = &[
     "SF Mono",
@@ -18,6 +19,17 @@ static MONO: OnceLock<String> = OnceLock::new();
 pub fn mono() -> &'static str {
     MONO.get_or_init(|| pick(MONO_CANDIDATES, "monospace"))
         .as_str()
+}
+
+// Char advance for the picked monospace face at `size`, measured via gpui's
+// text system. Falls back to ~7.2 px on resolve/measure failure (close to SF
+// Mono / Menlo at 12 px). Cheap to call per render; the platform font system
+// caches the lookup.
+pub fn mono_advance(cx: &App, size: Pixels) -> Pixels {
+    let font_id = cx.text_system().resolve_font(&font(mono()));
+    cx.text_system()
+        .ch_advance(font_id, size)
+        .unwrap_or(px(7.2))
 }
 
 fn pick(candidates: &[&str], fallback: &str) -> String {
